test: make HOME isolation reachable from integration tests (#588)
An integration test replaced the developer's real ~/.minutes/search.db with two
fixture rows pointing into a TempDir. Nothing in the test was wrong: search.db
and graph.db are resolved from the home directory rather than from Config, so
isolating output_dir does not reach them.

The isolation pattern already existed but was unreachable from tests/: the lock
was #[cfg(test)] pub(crate), and HomeOverride was duplicated privately inside
two unit-test modules. Integration tests link against the crate as an external
consumer and cannot see either. So the one test that touched global state was
the one that could not isolate it.

Adds a doc-hidden test_support module with a single HomeOverride, the shared
lock, and with_temp_home, compiled unconditionally so tests/ can use it.
Replaces both private copies and updates the 14 callers of the old lock helper.
Wraps the offending test in with_temp_home.

Verified empirically both ways: with a sentinel row in a real search.db, the
unfixed test replaces it with fixture rows pointing into /tmp; with the fix the
sentinel survives untouched.

Reported by @rymalia. Fixes #588
