+++ crates/core/src/events.rs
@@ -1309,7 +1309,7 @@ mod tests {
     use tempfile::TempDir;
 
     fn with_temp_home<T>(f: impl FnOnce(&TempDir) -> T) -> T {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         let original_home = std::env::var_os("HOME");
         let original_userprofile = std::env::var_os("USERPROFILE");
