+++ crates/core/src/live_transcript.rs
@@ -2904,7 +2904,7 @@ mod tests {
     use tempfile::{tempdir, NamedTempFile};
 
     fn with_temp_home<T>(f: impl FnOnce() -> T) -> T {
-        let _lock = crate::test_home_env_lock();
+        let _lock = crate::test_support::home_env_lock();
         let dir = tempfile::tempdir().unwrap();
         let original_home = std::env::var_os("HOME");
         #[cfg(windows)]
