+++ crates/core/src/search.rs
@@ -1456,7 +1456,7 @@ mod tests {
         // Search hits the HOME-derived sqlite index; serialize with the
         // crate HOME-env lock or a concurrently HOME-swapping test yanks
         // the index mid-write (sqlite disk I/O error).
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1485,7 +1485,7 @@ mod tests {
 
     #[test]
     fn search_returns_empty_for_no_match() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1513,7 +1513,7 @@ mod tests {
 
     #[test]
     fn search_is_case_insensitive() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1541,7 +1541,7 @@ mod tests {
 
     #[test]
     fn search_expands_vocabulary_aliases_with_provenance() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let home = TempDir::new().unwrap();
         unsafe {
             std::env::set_var("HOME", home.path());
@@ -1589,7 +1589,7 @@ mod tests {
         // Search hits the HOME-derived sqlite index; serialize with the
         // crate HOME-env lock or a concurrently HOME-swapping test yanks
         // the index mid-write (sqlite disk I/O error).
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1630,7 +1630,7 @@ mod tests {
 
     #[test]
     fn search_empty_directory() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         let config = Config {
             output_dir: dir.path().to_path_buf(),
@@ -1652,7 +1652,7 @@ mod tests {
 
     #[test]
     fn split_frontmatter_works() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let content = "---\ntitle: Test\ndate: 2026-03-17\n---\n\nBody text here.";
         let (fm, body) = split_frontmatter(content);
         assert!(fm.contains("title: Test"));
@@ -1661,7 +1661,7 @@ mod tests {
 
     #[test]
     fn extract_field_finds_value() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let fm = "title: My Meeting\ndate: 2026-03-17\ntype: meeting";
         assert_eq!(extract_field(fm, "title"), Some("My Meeting".into()));
         assert_eq!(extract_field(fm, "type"), Some("meeting".into()));
@@ -1670,7 +1670,7 @@ mod tests {
 
     #[test]
     fn search_intents_returns_matching_structured_records() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1708,7 +1708,7 @@ mod tests {
 
     #[test]
     fn search_intents_filters_by_kind_and_owner() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1741,7 +1741,7 @@ mod tests {
 
     #[test]
     fn search_intents_filters_owner_through_speaker_overlay() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         let meeting = dir.path().join("2026-03-17-test.md");
         create_test_file(
@@ -1787,7 +1787,7 @@ mod tests {
 
     #[test]
     fn search_intents_filter_by_recorded_by() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1835,7 +1835,7 @@ mod tests {
 
     #[test]
     fn consistency_report_flags_conflicts_and_stale_commitments() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1896,7 +1896,7 @@ mod tests {
 
     #[test]
     fn consistency_report_resolves_stale_owner_through_speaker_overlay() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         let meeting = dir.path().join("2020-03-01-a.md");
         create_test_file(
@@ -1931,7 +1931,7 @@ mod tests {
 
     #[test]
     fn consistency_report_marks_conflict_resolved_when_supersedes_is_set() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1961,7 +1961,7 @@ mod tests {
 
     #[test]
     fn consistency_report_leaves_resolution_none_without_supersedes() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -1987,7 +1987,7 @@ mod tests {
 
     #[test]
     fn consistency_report_does_not_mark_resolution_when_other_conflicts_remain() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -2019,7 +2019,7 @@ mod tests {
 
     #[test]
     fn consistency_report_requires_supersedes_to_reference_the_prior_decision() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -2044,7 +2044,7 @@ mod tests {
 
     #[test]
     fn consistency_report_ignores_near_duplicate_decisions() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -2068,7 +2068,7 @@ mod tests {
 
     #[test]
     fn person_profile_aggregates_recent_meetings_topics_and_open_intents() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -2099,7 +2099,7 @@ mod tests {
 
     #[test]
     fn person_profile_matches_linked_people_entities() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -2119,7 +2119,7 @@ mod tests {
 
     #[test]
     fn cross_meeting_research_collects_decisions_intents_and_meetings() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -2160,7 +2160,7 @@ mod tests {
 
     #[test]
     fn find_open_actions_parses_frontmatter() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         create_test_file(
             dir.path(),
@@ -2198,7 +2198,7 @@ mod tests {
 
     #[test]
     fn search_excludes_restricted_meetings_by_default() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = restricted_test_dir();
         let config = Config {
             output_dir: dir.path().to_path_buf(),
@@ -2226,7 +2226,7 @@ mod tests {
 
     #[test]
     fn search_intents_excludes_restricted_meetings_by_default() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = restricted_test_dir();
         let config = Config {
             output_dir: dir.path().to_path_buf(),
@@ -2252,7 +2252,7 @@ mod tests {
 
     #[test]
     fn find_open_actions_excludes_restricted_meetings_by_default() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = restricted_test_dir();
         let config = Config {
             output_dir: dir.path().to_path_buf(),
@@ -2272,7 +2272,7 @@ mod tests {
 
     #[test]
     fn cross_meeting_research_excludes_restricted_meetings_by_default() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = restricted_test_dir();
         let config = Config {
             output_dir: dir.path().to_path_buf(),
@@ -2306,7 +2306,7 @@ mod tests {
 
     #[test]
     fn person_profile_always_excludes_restricted_meetings() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = restricted_test_dir();
         let config = Config {
             output_dir: dir.path().to_path_buf(),
@@ -2323,7 +2323,7 @@ mod tests {
 
     #[test]
     fn consistency_report_always_excludes_restricted_meetings() {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = restricted_test_dir();
         let config = Config {
             output_dir: dir.path().to_path_buf(),
