+++ crates/core/src/sensitive.rs
@@ -402,7 +402,7 @@ mod tests {
     use tempfile::TempDir;
 
     fn with_temp_home<T>(f: impl FnOnce(&Config) -> T) -> T {
-        let _guard = crate::test_home_env_lock();
+        let _guard = crate::test_support::home_env_lock();
         let dir = TempDir::new().unwrap();
         let meetings = dir.path().join("meetings");
         let previous_home = std::env::var_os("HOME");
