+++ site/app/docs/errors/data.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-07-28",
+  "generatedAt": "2026-07-29",
   "visibleCount": 67,
   "hiddenCount": 17,
   "groups": [
